requirements.txt
# ==================== Web Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pydantic==2.5.0
pydantic-settings==2.1.0

# ==================== Database ====================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
pgvector==0.2.4  # PostgreSQLベクトル検索拡張

# ==================== OCR (Optical Character Recognition) ====================
pytesseract==0.3.10
# Tesseract本体は別途OSレベルでインストール必要
# macOS: brew install tesseract tesseract-lang
# Ubuntu: apt-get install tesseract-ocr tesseract-ocr-jpn

# ==================== Image Processing ====================
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2

# ==================== Auto Capture ====================
pyautogui==0.9.54
selenium==4.15.2
webdriver-manager==4.0.1

# ==================== RAG (Retrieval Augmented Generation) ====================
langchain==0.0.340
langchain-community==0.0.6
langchain-anthropic==0.0.2
langchain-openai==0.0.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
# GPU版を使う場合は faiss-cpu を faiss-gpu に変更

# ==================== LLM API ====================
anthropic==0.7.4
openai==1.3.7

# ==================== Task Queue ====================
celery==5.3.4
celery[redis]==5.3.4

# ==================== Rate Limiting ====================
slowapi==0.1.9

# ==================== Authentication & Security ====================
python-jose[cryptography]==3.3.0  # JWT token handling
passlib[bcrypt]==1.7.4  # Password hashing

# ==================== File Upload ====================
python-multipart==0.0.6  # FastAPI file upload

# ==================== Configuration ====================
python-dotenv==1.0.0

# ==================== Utilities ====================
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
httpx==0.25.2

# ==================== Monitoring & Logging ====================
python-json-logger==2.0.7

# ==================== Testing (Development) ====================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI

# ==================== Code Quality (Development) ====================
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# ==================== Documentation (Development) ====================
mkdocs==1.5.3
mkdocs-material==9.4.14
